requirements.txt
accelerate>=0.33.0
datasets
einops>=0.8.0
einx>=0.3.0
ema_pytorch>=0.5.2
faster_whisper
funasr
jieba
jiwer
librosa
matplotlib
pypinyin
safetensors
# torch>=2.0
# torchaudio>=2.3.0
torchdiffeq
tqdm>=4.65.0
transformers
vocos
wandb
x_transformers>=1.31.14
zhconv
zhon
gradio
cached-path